+++ docker-compose.yml
@@ -5,37 +5,55 @@ services:
   es-client:
     image: es-client:0.1.1
     build:
-      context: . # really only needs es.Dockerfile
+      context: .
       dockerfile: es.Dockerfile
-  # image: name:tag
     ports:
     - "127.0.0.1:9200:9200"
-    networks:
-      default:
-        ipv4_address: 172.19.0.3
+    # networks:
+    #   default:
+    #     ipv4_address: 172.19.0.3
+    healthcheck: # https://docs.docker.com/compose/compose-file/05-services/#healthcheck
+      test: curl -s http://host.docker.internal:9200 >/dev/null || exit 1
+      interval: 10s
+      timeout: 10s
+      retries: 10
+      start_period: 1m30s
+      # start_interval: 10s
 
 
-  py-peline:
-    image: py-peline:0.1.1
-    build:
-      context: . # needs app/ data/ and py.Dockerfile
-      dockerfile: py.Dockerfile
-  # image: name:tag
-    ports:
-      - "127.0.0.1:8888:8888"
-
-
-# TODO: branch out notebook from pipeline
-  # notebook:
-  #   image: notebook:0.1.1
+  # py-peline:
+  #   image: py-peline:0.1.1
   #   build:
   #     context: . 
-  #     dockerfile: jn.Dockerfile
-  # # image: name:tag
+  #     dockerfile: py.Dockerfile
+  #   ports:
+  #     - "127.0.0.1:8888:8888"
+  #   volumes:
+  #     - type: bind
+  #       source: ./src/pipeline
+  #       target: /app/src/pipeline
 
 
-networks:
-  default:
-    ipam:
-      config:
-        - subnet: 172.19.0.0/16
+
+  gui-search:
+    image: flask:0.1.1
+    build:
+      context: .
+      dockerfile: fl.Dockerfile
+    ports:
+      - "127.0.0.1:5001:5001"
+    depends_on:
+      es-client:
+        condition: service_healthy # https://docs.docker.com/compose/startup-order/
+        restart: true
+    volumes:
+       - type: bind
+         source: ./src
+         target: /app/src
+
+
+# networks:
+#   default:
+#     ipam:
+#       config:
+#         - subnet: 172.19.0.0/16
