+++ src/gui-search/.flaskenv
@@ -1,0 +1,3 @@
+FLASK_DEBUG=1
+FLASK_RUN_PORT=5001
+FLASK_RUN_HOST="0.0.0.0"
