+++ src/gui-search/README.md
@@ -1,0 +1,6 @@
+# Elasticsearch Search Tutorial
+
+This directory contains a starter Flask project used in the Search tutorial.
+
+https://github.com/elastic/elasticsearch-labs/raw/main/example-apps/search-tutorial/search-tutorial-starter.zip
+
