+++ src/gui-search/requirements.txt
@@ -1,0 +1,43 @@
+elasticsearch==8.12.1
+
+#
+# This file is autogenerated by pip-compile with Python 3.11
+# by the following command:
+#
+#    pip-compile requirements.in
+#
+blinker==1.6.3
+    # via flask
+build==1.0.3
+    # via pip-tools
+click==8.1.7
+    # via
+    #   flask
+    #   pip-tools
+flask==3.0.0
+    # via -r requirements.in
+itsdangerous==2.1.2
+    # via flask
+jinja2==3.1.2
+    # via flask
+markupsafe==2.1.3
+    # via
+    #   jinja2
+    #   werkzeug
+packaging==23.2
+    # via build
+pip-tools==7.3.0
+    # via -r requirements.in
+pyproject-hooks==1.0.0
+    # via build
+python-dotenv==1.0.0
+    # via -r requirements.in
+werkzeug==3.0.1
+    # via flask
+wheel==0.41.3
+    # via pip-tools
+
+# The following packages are considered to be unsafe in a requirements file:
+# pip
+# setuptools
+
