+++ src/gui-search/templates/document.html
@@ -1,0 +1,9 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <p><a href="javascript:history.back(1)" class="btn btn-primary">← Back</a></p>
+    <h2>{{ title }}</h2>
+    {% for paragraph in paragraphs %}
+        <p>{{ paragraph }}</p>
+    {% endfor %}
+{% endblock %}
